+++ examples/encode_decode_simple/encode_decode_simple.c
@@ -46,13 +46,11 @@ int main()
 
     kodo_factory_t encoder_factory =
         kodo_new_encoder_factory(code_type, finite_field,
-                                 max_symbols, max_symbol_size,
-                                 kodo_trace_disabled);
+                                 max_symbols, max_symbol_size);
 
     kodo_factory_t decoder_factory =
         kodo_new_decoder_factory(code_type, finite_field,
-                                 max_symbols, max_symbol_size,
-                                 kodo_trace_enabled);
+                                 max_symbols, max_symbol_size);
 
     // If we wanted to build an encoder of decoder with a smaller number of
     // symbols or a different symbol size, then this can be adjusted using the
@@ -76,7 +74,8 @@ int main()
     for(; i < block_size; ++i)
         data_in[i] = rand() % 256;
 
-    kodo_set_symbols(encoder, data_in, block_size);
+    kodo_set_const_symbols(encoder, data_in, block_size);
+    kodo_set_mutable_symbols(decoder, data_out, block_size);
 
     // Most of the network coding algorithms supports a mode of operation
     // which is known as systematic coding. This basically means that
@@ -107,11 +106,8 @@ int main()
     //    kodo_set_systematic_off(encoder);
     // }
 
-    // Install a custom trace function for the decoder (if tracing is enabled)
-    if (kodo_has_set_trace_callback(decoder))
-    {
-        kodo_set_trace_callback(decoder, trace_callback, NULL);
-    }
+    // Install a custom trace function for the decoder
+    kodo_set_trace_callback(decoder, trace_callback, NULL);
 
     while (!kodo_is_complete(decoder))
     {
@@ -127,8 +123,6 @@ int main()
         printf("Payload processed by decoder, current rank = %d\n",
                kodo_rank(decoder));
     }
-
-    kodo_copy_from_symbols(decoder, data_out, block_size);
 
     if (memcmp(data_in, data_out, block_size) == 0)
     {