+++ examples/sliding_window/sliding_window.c
@@ -48,13 +48,11 @@ int main()
     // The factories are used to build actual encoders/decoders
     kodo_factory_t encoder_factory =
         kodo_new_encoder_factory(code_type, finite_field,
-                                 max_symbols, max_symbol_size,
-                                 kodo_trace_disabled);
+                                 max_symbols, max_symbol_size);
 
     kodo_factory_t decoder_factory =
         kodo_new_decoder_factory(code_type, finite_field,
-                                 max_symbols, max_symbol_size,
-                                 kodo_trace_enabled);
+                                 max_symbols, max_symbol_size);
 
     // If we wanted to build an encoder or decoder with a smaller number of
     // symbols or a different symbol size, then this can be adjusted using the
@@ -77,6 +75,9 @@ int main()
     uint32_t block_size = kodo_block_size(encoder);
     uint8_t* data_in = (uint8_t*) malloc(block_size);
 
+    uint8_t* data_out = (uint8_t*) malloc(block_size);
+    kodo_set_mutable_symbols(decoder, data_out, block_size);
+
     uint8_t feedback_size = (uint8_t) kodo_feedback_size(encoder);
     uint8_t* feedback = (uint8_t*) malloc(feedback_size);
 
@@ -88,11 +89,8 @@ int main()
         data_in[i] = rand() % 256;
     }
 
-    // Install a custom trace function for the decoder (if tracing is enabled)
-    if (kodo_has_set_trace_callback(decoder))
-    {
-        kodo_set_trace_callback(decoder, trace_callback, NULL);
-    }
+    // Install a custom trace function for the decoder
+    kodo_set_trace_callback(decoder, trace_callback, NULL);
 
     while (!kodo_is_complete(decoder))
     {
@@ -101,7 +99,7 @@ int main()
         {
             uint32_t rank = kodo_rank(encoder);
             uint8_t* symbol = data_in + (rank * max_symbol_size);
-            kodo_set_symbol(encoder, rank, symbol, max_symbol_size);
+            kodo_set_const_symbol(encoder, rank, symbol, max_symbol_size);
             printf("Symbol %d added to the encoder\n", rank);
         }
 
@@ -141,9 +139,6 @@ int main()
         printf("Received feedback from decoder\n\n");
         kodo_read_feedback(encoder, feedback);
     }
-
-    uint8_t* data_out = (uint8_t*) malloc(kodo_block_size(decoder));
-    kodo_copy_from_symbols(decoder, data_out, kodo_block_size(decoder));
 
     if (memcmp(data_in, data_out, block_size) == 0)
     {