+++ src/kodoc/decoder_wrapper.hpp
@@ -51,17 +51,17 @@ namespace kodo
             return m_coder->is_complete();
         }
 
-        virtual void copy_from_symbols(uint8_t* data, uint32_t size) const
+        virtual void set_mutable_symbols(uint8_t* data, uint32_t size)
         {
-            auto storage = sak::mutable_storage(data, size);
-            m_coder->copy_from_symbols(storage);
+            sak::mutable_storage storage = sak::storage(data, size);
+            m_coder->set_mutable_symbols(storage);
         }
 
-        virtual void copy_from_symbol(uint32_t index, uint8_t* data,
-                                      uint32_t size) const
+        virtual void set_mutable_symbol(
+            uint32_t index, uint8_t* data, uint32_t size)
         {
-            auto storage = sak::mutable_storage(data, size);
-            m_coder->copy_from_symbol(index, storage);
+            sak::mutable_storage storage = sak::storage(data, size);
+            m_coder->set_mutable_symbol(index, storage);
         }
 
         virtual bool has_partial_decoding_tracker() const