+++ src/kodoc/kodo_encoder.hpp
@@ -22,6 +22,11 @@ extern "C"
         virtual void set_systematic_off() = 0;
         virtual void read_feedback(uint8_t* feedback) = 0;
 
+        virtual void set_const_symbols(uint8_t* data, uint32_t size) = 0;
+
+        virtual void set_const_symbol(uint32_t index, uint8_t* data,
+            uint32_t size) = 0;
+
         virtual uint32_t write_symbol(uint8_t* symbol_data,
             uint8_t* coefficients) = 0;
         virtual uint32_t write_uncoded_symbol(uint8_t* symbol_data,