+++ test/src/test_full_vector_codes.cpp
@@ -15,10 +15,7 @@ TEST(test_full_vector_codes, basic_api)
     uint32_t max_symbols = rand_symbols();
     uint32_t max_symbol_size = rand_symbol_size();
 
-    bool shallow_storage = false;
-
-    test_basic_api(kodo_full_vector, max_symbols, max_symbol_size,
-        shallow_storage);
+    test_basic_api(kodo_full_vector, max_symbols, max_symbol_size);
 }
 
 TEST(test_full_vector_codes, sparse_api)
@@ -27,31 +24,6 @@ TEST(test_full_vector_codes, sparse_api)
     uint32_t max_symbols = rand_symbols() + 10;
     uint32_t max_symbol_size = rand_symbol_size();
 
-    bool shallow_storage = false;
-
     test_basic_api(kodo_sparse_full_vector, kodo_full_vector, max_symbols,
-        max_symbol_size, shallow_storage);
-}
-
-TEST(test_full_vector_codes, shallow_api)
-{
-    uint32_t max_symbols = rand_symbols();
-    uint32_t max_symbol_size = rand_symbol_size();
-
-    bool shallow_storage = true;
-
-    test_basic_api(kodo_full_vector, max_symbols, max_symbol_size,
-        shallow_storage);
-}
-
-TEST(test_full_vector_codes, sparse_shallow_api)
-{
-    // Make sure that the decoding can complete with a lower density
-    uint32_t max_symbols = rand_symbols() + 10;
-    uint32_t max_symbol_size = rand_symbol_size();
-
-    bool shallow_storage = true;
-
-    test_basic_api(kodo_sparse_full_vector, kodo_full_vector, max_symbols,
-        max_symbol_size, shallow_storage);
+                   max_symbol_size);
 }