+++ test/src/test_helper.hpp
@@ -36,13 +36,12 @@ inline uint32_t rand_symbol_size(uint32_t max_symbol_size = 1600)
 }
 
 using test_function = std::function<
-    void (uint32_t, uint32_t, int32_t, int32_t, int32_t)>;
+    void (uint32_t, uint32_t, int32_t, int32_t)>;
 
 inline void test_combinations(
     test_function coder_test,
     uint32_t max_symbols,
-    uint32_t max_symbol_size,
-    bool trace_enabled)
+    uint32_t max_symbol_size)
 {
     SCOPED_TRACE(testing::Message() << "symbols = " << max_symbols);
     SCOPED_TRACE(testing::Message() << "symbol_size = " << max_symbol_size);
@@ -52,6 +51,8 @@ inline void test_combinations(
         kodo_full_vector,
         kodo_on_the_fly,
         kodo_sliding_window,
+        kodo_seed,
+        kodo_sparse_seed,
         kodo_perpetual
     };
 
@@ -59,8 +60,7 @@ inline void test_combinations(
     {
         kodo_binary,
         kodo_binary4,
-        kodo_binary8,
-        kodo_binary16
+        kodo_binary8
     };
 
     for (auto& code : code_types)
@@ -69,8 +69,7 @@ inline void test_combinations(
         {
             if (coder_test)
             {
-                coder_test(max_symbols, max_symbol_size,
-                           code, field, trace_enabled);
+                coder_test(max_symbols, max_symbol_size, code, field);
             }
         }
     }